+++ app/src/main/java/nextstep/signup/component/SignUpTextField.kt
@@ -6,13 +6,18 @@ import androidx.compose.material3.TextFieldColors
 import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
+import nextstep.signup.R
+import nextstep.signup.screen.SignUpTextFieldType
 import nextstep.signup.ui.theme.Blue50
 import nextstep.signup.ui.theme.Gray10
 import nextstep.signup.ui.theme.Gray40
 import nextstep.signup.ui.theme.Red40
 import nextstep.signup.ui.theme.SignupTheme
+import nextstep.signup.util.validation.ValidationErrorType
+import nextstep.signup.util.validation.ValidationResult
 
 @Composable
 fun SignUpTextField(
@@ -52,6 +57,33 @@ fun SignUpTextField(
     )
 }
 
+
+@Composable
+fun signUpSupportingTextStringResource(
+    signUpTextFieldType: SignUpTextFieldType,
+    validationResult: ValidationResult
+): String? {
+    return if (validationResult is ValidationResult.ValidationError) {
+        when (signUpTextFieldType) {
+            SignUpTextFieldType.UserName -> {
+                when (validationResult.type) {
+                    ValidationErrorType.LengthError ->
+                        stringResource(id = R.string.sign_up_user_name_length_error)
+
+                    ValidationErrorType.RegexError ->
+                        stringResource(id = R.string.sign_up_user_name_regex_error)
+                }
+            }
+            SignUpTextFieldType.Email -> {
+                if(validationResult.type == ValidationErrorType.RegexError) {
+                    stringResource(id = R.string.sign_up_email_regex_error)
+                }
+                else null
+            }
+        }
+
+    } else null
+}
 
 @Preview(showBackground = true)
 @Composable
