+++ app/src/main/java/nextstep/signup/screen/SignUpScreen.kt
@@ -1,6 +1,5 @@
 package nextstep.signup.screen
 
-import android.util.Log
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -25,15 +24,14 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import nextstep.signup.R
+import nextstep.signup.component.EmailTextField
 import nextstep.signup.component.SignUpTextField
 import nextstep.signup.component.UserNameTextField
 import nextstep.signup.ui.theme.Blue50
 import nextstep.signup.ui.theme.SignupTheme
-import nextstep.signup.util.validation.ValidationErrorType
-import nextstep.signup.util.validation.ValidationResult
 
 enum class SignUpTextFieldType {
-    UserName
+    UserName, Email
 }
 
 @Composable
@@ -66,13 +64,12 @@ fun SignUpScreen(
                     userName = value
                 },
             )
-            SignUpTextField(
+            EmailTextField(
                 modifier = Modifier.fillMaxWidth(),
                 text = email,
                 onValueChange = { value ->
                     email = value
-                },
-                labelText = stringResource(id = R.string.sign_up_email_label),
+                }
             )
             SignUpTextField(
                 modifier = Modifier.fillMaxWidth(),
@@ -109,28 +106,6 @@ fun SignUpScreen(
             }
         }
     }
-}
-
-@Composable
-fun signUpSupportingTextStringResource(
-    signUpTextFieldType: SignUpTextFieldType,
-    validationResult: ValidationResult
-): String? {
-    return if (validationResult is ValidationResult.ValidationError) {
-        when (signUpTextFieldType) {
-            SignUpTextFieldType.UserName -> {
-                when (validationResult.type) {
-                    ValidationErrorType.LengthError ->
-                        stringResource(id = R.string.sign_up_user_name_length_error)
-
-                    ValidationErrorType.RegexError ->
-                        stringResource(id = R.string.sign_up_user_name_regex_error)
-                }
-            }
-        }
-
-    } else null
-
 }
 
 @Preview(showBackground = true)
