+++ app/src/main/res/values/strings.xml
@@ -9,4 +9,5 @@
 
     <string name="sign_up_user_name_length_error">이름은 2~5자여야 합니다.</string>
     <string name="sign_up_user_name_regex_error">이름에는 숫자나 기호가 포함될 수 없습니다.</string>
+    <string name="sign_up_email_regex_error">이메일 형식이 올바르지 않습니다.</string>
 </resources>
